billing settings - moved methods to SettingStorage
