+++ src/storage/mongodb/BillingStorage.ts
@@ -1,5 +1,4 @@
 import { BillingAdditionalData, BillingInvoice, BillingInvoiceStatus, BillingSessionData } from '../../types/Billing';
-import { BillingSettings, BillingSettingsType, SettingDB } from '../../types/Setting';
 import global, { FilterParams } from '../../types/GlobalType';
 
 import Constants from '../../utils/Constants';
@@ -8,8 +7,6 @@ import DatabaseUtils from './DatabaseUtils';
 import DbParams from '../../types/database/DbParams';
 import Logging from '../../utils/Logging';
 import { ObjectID } from 'mongodb';
-import SettingStorage from './SettingStorage';
-import TenantComponents from '../../types/TenantComponents';
 import Utils from '../../utils/Utils';
 
 const MODULE_NAME = 'BillingStorage';
@@ -236,59 +233,5 @@ export default class BillingStorage {
       .findOneAndDelete({ 'invoiceID': id });
     // Debug
     await Logging.traceEnd(tenantID, MODULE_NAME, 'deleteInvoice', uniqueTimerID, { id });
-  }
-
-  public static async getBillingSetting(tenantID: string): Promise<BillingSettings> {
-    // Get BILLING Settings by Identifier
-    const setting = await SettingStorage.getSettingByIdentifier(tenantID, TenantComponents.BILLING);
-    if (setting) {
-      return BillingStorage.convertToBillingSettings(setting);
-    }
-    return null;
-  }
-
-  public static async saveBillingSetting(tenantID: string, billingSettings: BillingSettings): Promise<string> {
-    const { id, identifier, sensitiveData, backupSensitiveData, category } = billingSettings;
-    const { createdBy, createdOn, lastChangedBy, lastChangedOn } = billingSettings;
-    const { type, billing, stripe } = billingSettings;
-    const setting: SettingDB = {
-      id, identifier, sensitiveData, backupSensitiveData,
-      content: {
-        type,
-        billing,
-      },
-      category, createdBy, createdOn, lastChangedBy, lastChangedOn,
-    };
-    if (billingSettings.type === BillingSettingsType.STRIPE) {
-      setting.sensitiveData = [ 'content.stripe.secretKey' ];
-      setting.content.stripe = stripe;
-    }
-    return SettingStorage.saveSettings(tenantID, setting);
-  }
-
-  private static convertToBillingSettings(setting: SettingDB): BillingSettings {
-    const { id, backupSensitiveData, category } = setting;
-    const { createdBy, createdOn, lastChangedBy, lastChangedOn } = setting;
-    const { content } = setting;
-    const billingSettings: BillingSettings = {
-      id,
-      identifier: TenantComponents.BILLING,
-      type: content.type as BillingSettingsType,
-      backupSensitiveData,
-      billing: content.billing,
-      category,
-      createdBy,
-      createdOn,
-      lastChangedBy,
-      lastChangedOn,
-    };
-    switch (content.type) {
-      // Only STRIPE so far
-      case BillingSettingsType.STRIPE:
-        billingSettings.stripe = content.stripe;
-        billingSettings.sensitiveData = [ 'stripe.secretKey' ];
-        break;
-    }
-    return billingSettings;
   }
 }
